locking/mutex: split out mutex_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performance

Trimming down sched.h dependencies: we don't want to include more than
the base types.

Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Will Deacon <will@kernel.org>
Cc: Waiman Long <longman@redhat.com>
Cc: Boqun Feng <boqun.feng@gmail.com>
(cherry picked from commit d84f317915172c6511fd6c14ea3f70c9d67fdf67)
